+++ CHANGELOG.md
@@ -2,4 +2,10 @@
 Below are major changes for each version Release. For detailed information,
 see the list of commits from the last version or use `git log`.
 
-## 0.0.0
+## 0.0.1
+
+- Initial development and gladier setup
+
+## 0.0.2
+
+- Refactor and clean